+++ packages/agent/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/agent [1.0.0-alpha.6](https://github.com/anolilab/lunora/compare/%40lunora%2Fagent%401.0.0-alpha.5...%40lunora%2Fagent%401.0.0-alpha.6) (2026-07-21)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.29
+
 ## @lunora/agent [1.0.0-alpha.5](https://github.com/anolilab/lunora/compare/%40lunora%2Fagent%401.0.0-alpha.4...%40lunora%2Fagent%401.0.0-alpha.5) (2026-07-20)