+++ packages/agent/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/agent [1.0.0-alpha.7](https://github.com/anolilab/lunora/compare/%40lunora%2Fagent%401.0.0-alpha.6...%40lunora%2Fagent%401.0.0-alpha.7) (2026-07-22)
+
+
+### Dependencies
+
+* **@lunora/ai:** upgraded to 1.0.0-alpha.18
+
 ## @lunora/agent [1.0.0-alpha.6](https://github.com/anolilab/lunora/compare/%40lunora%2Fagent%401.0.0-alpha.5...%40lunora%2Fagent%401.0.0-alpha.6) (2026-07-21)